AreaMap part 3: Refactor any code which calculates an axon index in preparation for new system. Lots of bismit.cl cleanup and reorganization.

+++ src/cortex.rs
@@ -105,7 +105,7 @@ impl Cortex {
 		self.area_mut(ao_name).axns.states.read();
 
 		let (out_start_ao, out_end_ao) = self.area(ao_name).mcols.axn_output_range();
-		let out_slc_ao = &self.area(ao_name).axns.states.vec[out_start_ao..out_end_ao];
+		let out_slc_ao = &self.area(ao_name).axns.states.vec[out_start_ao as usize..out_end_ao as usize];
 
 		let cols = self.area(ao_name).dims.columns(); // DEBUG PURPOSES
 		println!("Area: '{}' - out_start_ao: {}, out_end_ao: {}, cols: {}", ao_name, out_start_ao, out_end_ao, cols);
@@ -170,7 +170,7 @@ impl Cortex {
 
 		for slc in axn_slcs { 
 			let buffer_offset = cmn::axn_idz_2d(slc, self.areas.get(area_name).expect(emsg).dims.columns(), region.hrz_demarc()) as usize;
-			//let buffer_offset = cmn::AXON_MARGIN_SIZE + (axn_slc as usize * self.cortical_area.axns.dims.width as usize);
+			//let buffer_offset = cmn::AXON_MAR__GIN_SIZE + (axn_slc as usize * self.cortical_area.axns.dims.width as usize);
 
 			//println!("##### write_vec(): {} offset: axn_idz_2d(axn_slc: {}, dims.columns(): {}, region.hrz_demarc(): {}): {}, vec.len(): {}", layer_target, slc, self.cortical_area.dims.columns(), region.hrz_demarc(), buffer_offset, vec.len());
